Refactor: Change import paths for interfaces
