+++ Src/Commands/Active/jwst.ts
@@ -8,7 +8,7 @@ import { type LCARSClient } from '../../Subsystems/Auxiliary/LCARSClient.js';
 import https from 'https';
 
 import Utility from '../../Subsystems/Utilities/SysUtils.js';
-import { type Command } from '../../Subsystems/Auxiliary/CommandInterface';
+import { type Command } from '../../Subsystems/Auxiliary/Interfaces/CommandInterface';
 
 // Functions
 const data = new SlashCommandBuilder()
